waitm-653: hide manual sync in avatar
